scripts/requirements.txt
# Requirements for utility scripts
# Install with: pip install -r scripts/requirements.txt

# LINE Bot SDK (same version as Cloud Functions)
line-bot-sdk==3.12.0

# Google Cloud Libraries
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0

# Image processing (for rich menu setup)
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0
